requirements.txt
alpaca-py>=0.38.0
yfinance>=0.2.36
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.0.0
scipy>=1.11.0
requests>=2.31.0
loguru>=0.7.0
